# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBTS API
  description: API for OpenBTS Backend
  version: 1.0.0
servers:
  - url: https://openbts.sakilabs.com/api/v1
    description: API v1 base path
...
paths:
...
  /stations/{station_id}/comments:
    get:
      tags: [Stations]
      summary: List comments for a station
      parameters:
        - $ref: "#/components/parameters/StationIdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/StationComment" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: [Stations]
      summary: Create comment (multipart/form-data with content and optional files)
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/StationIdPath"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                file: { type: string, format: binary, description: "Optional image attachment" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/StationComment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /submissions/{id}/photos:
    get:
      tags: [Submissions]
      summary: List photos for a submission
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/SubmissionPhoto" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      tags: [Submissions]
      summary: Upload photos for a submission (multipart/form-data)
      description: Accepts interleaved `notes` fields and `files` fields. Max 5 photos per submission, max 3 MB per file. Only the submission's owner can upload while status is pending.
      security:
        - ApiKeyAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                notes: { type: string, description: "Optional note for the corresponding file (max 100 chars)" }
                files: { type: string, format: binary }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, attachment_uuid, mime_type, createdAt]
                      properties:
                        id: { type: integer }
                        attachment_uuid: { type: string, format: uuid }
                        mime_type: { type: string }
                        createdAt: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API Key for authentication via X-Api-Key header
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Cookie-based authentication
  parameters:
...
    StationIdPath:
      name: station_id
      in: path
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Bad request (invalid parameters, validation error, or invalid query)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [BAD_REQUEST, VALIDATION_ERROR, INVALID_QUERY]
          example:
            errors:
              - code: BAD_REQUEST
                message: Invalid request
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [UNAUTHORIZED]
          example:
            errors:
              - code: UNAUTHORIZED
                message: Unauthorized
    Forbidden:
      description: Forbidden or insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [FORBIDDEN, INSUFFICIENT_PERMISSIONS]
          example:
            errors:
              - code: FORBIDDEN
                message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [NOT_FOUND]
          example:
            errors:
              - code: NOT_FOUND
                message: Resource not found
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [TOO_MANY_REQUESTS]
          example:
            errors:
              - code: TOO_MANY_REQUESTS
                message: Too many requests
...
    InternalServerError:
      description: Internal server error (e.g. failed to create/update/delete)
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiError"
              - type: object
                properties:
                  errors:
                    items:
                      properties:
                        code:
                          enum: [INTERNAL_SERVER_ERROR, FAILED_TO_CREATE, FAILED_TO_UPDATE, FAILED_TO_DELETE, AUTH_FAILURE]
          example:
            errors:
              - code: INTERNAL_SERVER_ERROR
                message: Internal server error
  schemas:
    ApiError:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                enum:
                  - INTERNAL_SERVER_ERROR
                  - BAD_REQUEST
                  - NOT_FOUND
                  - UNAUTHORIZED
                  - FORBIDDEN
                  - FAILED_TO_DELETE
                  - INVALID_QUERY
                  - FAILED_TO_UPDATE
                  - FAILED_TO_CREATE
                  - AUTH_FAILURE
                  - VALIDATION_ERROR
                  - ALREADY_LOGGED_IN
                  - INSUFFICIENT_PERMISSIONS
                  - TOO_MANY_REQUESTS
                  - DUPLICATE_ENTRY
              message: { type: string }
              details:
                type: array
                description: Only present for VALIDATION_ERROR
                items:
                  type: object
                  properties:
                    field: { type: string }
                    validationMessage: { type: string, nullable: true }
...
    PhotoAuthor:
      type: object
      nullable: true
      properties:
        uuid: { type: string, format: uuid }
        username: { type: string }
        name: { type: string }
...
    SubmissionPhoto:
      type: object
      required: [id, attachment_uuid, mime_type, note, createdAt, author]
      properties:
        id: { type: integer }
        attachment_uuid: { type: string, format: uuid }
        mime_type: { type: string }
        note: { type: string, maxLength: 100, nullable: true }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/PhotoAuthor" }
...
    StationComment:
      type: object
      properties:
        id: { type: integer }
        station_id: { type: integer }
        user_id: { type: string, format: uuid }
        content: { type: string }
        attachments: { type: array, items: { type: object, properties: { uuid: { type: string }, type: { type: string } } }, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...